crazyflie_benchmark/requirements.txt
# Core dependencies
numpy>=1.20.0
matplotlib>=3.4.0
pyyaml>=6.0
scipy>=1.7.0

# For real hardware
# cflib>=0.1.14  # Uncomment to install Crazyflie library

# For simulator
# requests>=2.27.0  # Uncomment to install requests library

# Optional dependencies
pandas>=1.3.0  # For data processing 